Created SproutEmail_DefaultMailerRecipientListRecipientRecord that will relate the email recipient record and the list record.
Modify EmailRecipientRecord to have a relation with the ListRecipientRecord.

Modify ListRecord to have a relation with the ListRecipientRecord.

Refactor the saving of email recipient that saves relation between email recipient and list record to the ListRecipientRecord

Modify the EmailRecipientElementType to join the new ListRecipientRecord and remove old attributes.

Modify models with the new attributes and added delete script for deleting the ListRecipientRecord after deletiting email and list records.